started implementing dx11 and restructure of gfx layer (at the moment only a red quad is being rendered) opengl is currently deactivated and needs to be adjusted to the new gfx layer added vulkan api (not working yet)

+++ Dawn/Dawn/src/Dawn/ResourceSystem/ResourceWatcher.cpp
@@ -39,7 +39,7 @@ namespace Dawn
 		do
 		{
 			bPending = ReadDirectoryChangesW(InData->DirHandle,
-				&buffer[0], buffer.size(),
+				&buffer[0], DWORD(buffer.size()),
 				TRUE, FILE_NOTIFY_CHANGE_FILE_NAME |
 				 FILE_NOTIFY_CHANGE_SIZE,
 				&dwBytes, &o, NULL);
